pyproject.toml
[tool.poetry]
name = "llama-workflow-groq"
version = "0.1.0"
description = "Advanced RAG with llamaworkflow with Groq"
authors = ["diegogs <diego.souza@me.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.11.19"
llama-index-llms-groq = "^0.2.0"
llama-index-embeddings-huggingface = "^0.3.1"
llama-index-utils-workflow = "^0.2.1"
pyvis = "^0.3.2"
python-dotenv = "^1.0.1"
nest-asyncio = "^1.6.0"
graphviz = "^0.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
